requirement.txt
fastapi
uvicorn
sqlmodel
alembic
bcrypt
python-jose
python-dotenv
passlib[bcrypt]
celery
